Update for Github Pages.

+++ deploy-to-gh-pages.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 set -e # exit with nonzero exit code if anything fails
 
+if [ "$TRAVIS_PULL_REQUEST" != "false" -o "$TRAVIS_BRANCH" != "master" ]; then exit 0; fi
+
 # clear and re-create the out directory
 rm -rf out || exit 0;
 mkdir out;
@@ -8,12 +10,15 @@ mkdir out;
 # run our compile script, discussed above
 jupyter nbconvert Lecture*.ipynb --to slides --output-dir ./out
 
+mkdir out/reveal.js
+cp -r reveal.js/ out/reveal.js/
+
 # go to the out directory and create a *new* Git repo
 cd out
 git init
 
 # inside this git repo we'll pretend to be a new user
-git config user.name "Travis CI"
+git config user.name "Non-Contradiction"
 git config user.email "lch34677@gmail.com"
 
 # The first and only commit to this new Git repo contains all the
